styling new library home; menubar/links
