+++ src/components/App.css
@@ -3,7 +3,7 @@
 .App-container {
     max-width: 1400px;
     margin: 0 auto;
-    padding: 0 10px;
+    padding: 0;
 }
 body {
     background-color: #101010;
@@ -52,7 +52,7 @@ span.MuiButton-root .MuiButton-label{
 } */
 
 .header-spacer {
-    height: 6em;
+    height: 3em;
     margin-bottom: 10px;
 }
 .nft-flex-layout {
@@ -92,3 +92,22 @@ hr.fullDivider {
 .center-align {
     text-align: center;
 }
+
+
+
+
+.footer {
+    background-color: #101010;
+    height: 4em;
+    width: 100%;
+    position: fixed;
+    bottom: 0;
+}
+    .footer .footer-logo-player {
+        height: 50px;
+        margin: 0 auto;
+        text-align: center;
+    }
+    .footer .footer-logo-player img {
+        height: 100%;
+    }
