+++ src/components/App.js
@@ -1,10 +1,11 @@
 import React, { Component } from 'react';
 import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
 import Header from'./Header';
-import DisplayTokens from'./DisplayTokens';
+import DisplayTokens from'./player/DisplayTokens';
 import UploadToIPFS from'./UploadToIPFS';
 import Styles from'./Styles';
 import TokenForm from'./TokenForm';
+import MusicHome from'./player/MusicHome';
 import Melomaniac from '../abis/Melomaniac.json'
 import './App.css';
 import Web3 from 'web3';
@@ -80,10 +81,13 @@ class App extends Component {
             <div className="header-spacer"></div>
           
             <Switch>
-              <Route path="/" exact component={() => <DisplayTokens />} />
+              <Route path="/" exact component={() => <MusicHome />} />
+              <Route path="displayTokens" exact component={() => <DisplayTokens />} />
               <Route path="/tokenForm" exact component={() => <TokenForm />} />
               <Route path="/upload" exact component={() => <UploadToIPFS />} />
+              {/* Experiments */}
               <Route path="/styles" exact component={() => <Styles />} />
+              <Route path="/musicHome" exact component={() => <MusicHome />} />
             </Switch>
           </Router>
