+++ src/components/Header.css
@@ -10,7 +10,7 @@ p{
     flex-flow: row wrap;
     /* This aligns items to the end line on main-axis */
     justify-content: space-between;
-    padding: 8px 10px;
+    padding: 12px 10px;
     position: fixed;
     top: 0;
     left: 0;
@@ -21,15 +21,17 @@ p{
 /* Medium screens */
 @media all and (max-width: 800px) {
     .navbar-ui {
-      /* When on medium sized screens, we center it by evenly distributing empty space around items */
-      justify-content: space-around;
+      /* justify-content: space-around; */
     }
   }
   /* Small screens */
 @media all and (max-width: 500px) {
     .navbar-ui {
       /* On small screens, we are no longer using row direction but column */
-      flex-direction: column;
+      /* flex-direction: column; */
+    }
+    .navbar-ui .App-logo {
+      display:none;
     }
   }
     .navbar-ui a {
