+++ src/components/player/MusicHome.css
@@ -1,0 +1,30 @@
+/* menu */
+.playerMenu {
+    width: 100%;
+}
+    .playerMenu .playerMenuList {
+        list-style:none;
+        margin: 2px 0;
+        padding: 2px;
+    }
+        .playerMenuList li {
+            padding: .8em;
+            font-size: 1.3em;
+            background-color: #21132C;
+            display: flex;
+            justify-content: space-between;
+            margin-bottom: 2px;
+        }
+    
+    .playerMenuList .menulist-icon.MuiSvgIcon-root {
+        margin-right:.8em;
+        flex-basis: 1.2em;
+    }
+    .playerMenuList li span {
+        flex-grow: 4;
+    }
+    .playerMenuList .angle-button.MuiSvgIcon-root {
+        align-self: flex-end;
+        margin-right: 0;
+        flex-basis: 2em;
+    }
