+++ src/components/player/MusicPreviewCard.css
@@ -1,0 +1,54 @@
+.card {
+    margin: 10px;
+    box-shadow: 1px 1px 38px 1px rgba(0,0,0,0.4);
+    -webkit-box-shadow: 1px 1px 38px 1px rgba(0,0,0,0.4);
+    -moz-box-shadow: 1px 1px 38px 1px rgba(0,0,0,0.4);
+}
+.card.nft-card {
+    padding:10px;
+}
+    .nft-card audio {
+        width: auto;
+        margin-bottom: 10px;
+    }
+    .nft-card .card-image {
+        overflow: hidden;
+        height: 200px;
+        width: 100%;
+        margin-bottom: 10px;
+    }
+        .nft-card .card-image img{
+            /* height: 200px; */
+            width: 100%;
+            height: 100%;
+            object-fit: cover;
+            
+        }
+    .nft-card h3,
+    .nft-card h5 {
+        color: #343434;
+    }
+
+.sample-section{
+    width: 100%;
+    margin: 0 auto;
+    max-width:1200px;
+}
+.token-price-container {
+    display: flex;
+    justify-content: space-between;
+}
+
+button {
+    width: 100%;
+    max-width: 260px;
+    box-shadow: 1px 1px 12px 1px rgba(0,0,0,0.6);
+    -webkit-box-shadow: 1px 1px 12px 1px rgba(0,0,0,0.6);
+    -moz-box-shadow: 1px 1px 12px 1px rgba(0,0,0,0.6);
+    border-radius: 2em;
+    border: none;
+    background-color:#8F00FF;
+    padding:10px 20px;
+    color: #ffffff;
+    margin: 0 auto;
+}
